backend/requirements.txt
a2a-sdk==0.2.7
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
argcomplete==3.6.2
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.4
billiard==4.2.1
boto3==1.38.34
botocore==1.38.34
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cohere==5.15.0
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
ecdsa==0.19.1
eval_type_backport==0.2.2
fasta2a==0.2.16
fastapi==0.115.12
fastavro==1.11.1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.0
google-auth==2.40.3
google-genai==1.19.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
groq==0.27.0
grpcio==1.73.0
grpcio-reflection==1.73.0
grpcio-tools==1.73.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
kombu==5.5.4
llama-cloud==0.1.26
llama-cloud-services==0.6.31
llama-index==0.12.41
llama-index-agent-openai==0.4.10
llama-index-cli==0.4.3
llama-index-core==0.12.41
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.6
llama-index-llms-openai==0.4.5
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.31
logfire-api==3.18.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.9.3
mdurl==0.1.2
mistralai==1.8.2
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.0
openai==1.86.0
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-ai==0.2.16
pydantic-ai-slim==0.2.16
pydantic-evals==0.2.16
pydantic-graph==0.2.16
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pypdf==5.6.0
pytest==8.0.0
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zope.event==5.0
zope.interface==7.2
